Patch 0.1.3: Refactored About Me from page -> grid -> component
